Refactor [#92] 리뷰 반영하여 ValidateService 생성 후 분리
