+++ morib/src/main/java/org/morib/server/domain/relationship/application/FetchRelationshipService.java
@@ -10,6 +10,6 @@ public interface FetchRelationshipService {
     List<Relationship> fetchConnectedRelationship(Long userId);
     List<Relationship> fetchUnconnectedRelationship(Long userId);
     Relationship fetchRelationshipByUserIdAndFriendId(Long userId, Long friendId, RelationLevel relationLevel);
-    void validateRelationshipByUserAndFriend(User user, User friend);
+
 
 }
